+++ README.md
@@ -1,11 +1,12 @@
 # Yellow Ketchapp
 
-This repository provides templates for a gRPC server written in Go and a simple GUI client written in Python. The server now exposes an `ImageService` API so clients can list and download image files.
+This repository provides templates for gRPC servers written in Go and C++ as well as a simple GUI client written in Python. The servers only implement a `Greeter` service.
 
-- `server/` contains the Go gRPC server template.
-- `client_pyside/` contains the Python GUI client template built with PySide6.
-- `client_qtwidget/` contains a Qt gRPC client implemented in C++ using QtProtobuf.
-- `client_qtquick/` contains a Qt Quick client implemented in QML using QtProtobuf.
+- `server/go/` contains the Go gRPC server template.
+- `server/cpp/` contains a C++ gRPC server example.
+- `client/pyside/` contains the Python GUI client template built with PySide6.
+- `client/qtwidget/` contains a Qt gRPC client implemented in C++ using QtProtobuf.
+- `client/qtquick/` contains a Qt Quick client implemented in QML using QtProtobuf.
 - `proto/` holds the shared Protocol Buffers definitions used by both components.
 
 Refer to the README files in each directory for details on building and running the components.
@@ -13,7 +14,7 @@ For the Python client specifically, make sure to install its requirements before
 generating the gRPC stubs:
 
 ```bash
- pip install -r client_pyside/requirements.txt
+ pip install -r client/pyside/requirements.txt
 ```
 
 ## Generating gRPC Code
@@ -22,15 +23,18 @@ Both components share the top-level `proto/` directory. You can generate the
 language-specific stubs from the repository root with the following commands:
 
 ```bash
-# Go stubs for the server
-protoc -I ./proto --go_out=./server --go-grpc_out=./server ./proto/imagestorage.proto
+# Go stubs for the Go server
+protoc -I ./proto --go_out=./server/go --go-grpc_out=./server/go ./proto/helloworld.proto
 
 # Python stubs for the PySide client
-python -m grpc_tools.protoc -I ./proto --python_out=./client_pyside --grpc_python_out=./client_pyside ./proto/imagestorage.proto
+python -m grpc_tools.protoc -I ./proto --python_out=./client/pyside --grpc_python_out=./client/pyside ./proto/helloworld.proto
 
 # Qt C++ stubs for the Qt client
-protoc -I ./proto --qt_out=./client_qtwidget --qt-grpc_out=./client_qtwidget ./proto/imagestorage.proto
-protoc -I ./proto --qt_out=./client_qtquick --qt-grpc_out=./client_qtquick ./proto/imagestorage.proto
+protoc -I ./proto --qt_out=./client/qtwidget --qt-grpc_out=./client/qtwidget ./proto/helloworld.proto
+protoc -I ./proto --qt_out=./client/qtquick --qt-grpc_out=./client/qtquick ./proto/helloworld.proto
+
+# C++ stubs for the C++ server
+protoc -I ./proto --cpp_out=./server/cpp --grpc_out=./server/cpp ./proto/helloworld.proto
 ```
 
-The Qt client is built using CMake. See `client_qtwidget/README.md` for details on configuring a Qt environment and compiling the application.
+The Qt client is built using CMake. See `client/qtwidget/README.md` for details on configuring a Qt environment and compiling the application.
