+++ client/pyside/README.md
@@ -1,8 +1,7 @@
 # PySide gRPC Client Template
 
 This directory contains a simple GUI application written in Python using the PySide6 toolkit that communicates with the gRPC server.
-In addition to sending a greeting, the client can now list images available on the server and download the selected file.
-
+This example simply sends a greeting using the `Greeter` service.
 ## Prerequisites
 
 - Python 3.11 or later
@@ -33,7 +32,7 @@ After installing the dependencies, generate the Python gRPC code from the shared
 `.proto` file by running the command below from the repository root:
 
 ```bash
-python -m grpc_tools.protoc -I ./proto --python_out=./client_pyside --grpc_python_out=./client_pyside ./proto/imagestorage.proto
+python -m grpc_tools.protoc -I ./proto --python_out=./client/pyside --grpc_python_out=./client/pyside ./proto/helloworld.proto
 ```
 
 ## Running the Application
