+++ client/pyside/requirements.txt
