+++ client/qtquick/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.16)
 project(client_qtquick LANGUAGES CXX)
 
-find_package(Qt6 REQUIRED COMPONENTS Core Quick Grpc Protobuf GrpcQuick ProtobufQuick ProtobufQtCoreTypes)
+find_package(Qt6 REQUIRED COMPONENTS Core Quick Grpc Protobuf GrpcQuick ProtobufQuick)
 
 qt_standard_project_setup()
 
@@ -21,28 +21,22 @@ qt_add_protobuf(client_qtquick_proto
     QML
     QML_URI ClientQtQuick.Proto
     PROTO_FILES
-        ${CMAKE_CURRENT_SOURCE_DIR}/../proto/imagestorage.proto
-    PROTO_INCLUDES
-        $<TARGET_PROPERTY:Qt6::ProtobufQtCoreTypes,QT_PROTO_INCLUDES>
+        ${CMAKE_CURRENT_SOURCE_DIR}/../../proto/helloworld.proto
 )
 
 qt_add_grpc(client_qtquick_proto CLIENT
-    PROTO_FILES
-        ${CMAKE_CURRENT_SOURCE_DIR}/../proto/imagestorage.proto
-    PROTO_INCLUDES
-        $<TARGET_PROPERTY:Qt6::ProtobufQtCoreTypes,QT_PROTO_INCLUDES>
     QML
+    PROTO_FILES
+        ${CMAKE_CURRENT_SOURCE_DIR}/../../proto/helloworld.proto
 )
 
 target_link_libraries(client_qtquick_proto PRIVATE
     Qt6::Protobuf
-    Qt6::ProtobufQtCoreTypes
     Qt6::Grpc
 )
 
 target_link_libraries(client_qtquick PRIVATE
     Qt6::Core
     Qt6::Quick
-    Qt6::ProtobufQtCoreTypes
     client_qtquick_proto
 )
