+++ client/qtquick/Main.cpp
@@ -1,11 +1,8 @@
 #include <QGuiApplication>
 #include <QQmlApplicationEngine>
-#include <QtProtobufQtCoreTypes>
 
 int main(int argc, char *argv[])
 {
-    QtProtobuf::registerProtobufQtCoreTypes();
-
     QGuiApplication app(argc, argv);
 
     QQmlApplicationEngine engine;
