+++ client/qtquick/README.md
@@ -2,11 +2,8 @@
 
 This directory provides a minimal QML application built with Qt Quick. The code demonstrates how QtGrpc and QtProtobuf can be used from QML to talk to the gRPC server.
 
-The example mirrors the features of the widgets client:
-
-- Send a greeting using the `Greeter` service
-- List images available on the server
-- Download the selected image
+The example mirrors the widgets client and simply sends a greeting using the
+`Greeter` service.
 
 ## Building
