+++ generate_proto.sh
@@ -4,13 +4,19 @@ set -e
 # Python 코드 생성
 python -m grpc_tools.protoc \
   -I=./proto \
-  --python_out=./client_pyside/proto \
-  --grpc_python_out=./client_pyside/proto \
-  ./proto/imagestorage.proto
+  --python_out=./client/pyside/proto \
+  --grpc_python_out=./client/pyside/proto \
+  ./proto/helloworld.proto
 
 # Go 코드 생성
 protoc \
   -I=./proto \
-  --go_out=. \
-  --go-grpc_out=. \
-  ./proto/imagestorage.proto
+  --go_out=./server/go \
+  --go-grpc_out=./server/go \
+  ./proto/helloworld.proto
+
+# C++ 코드 생성
+protoc \
+  -I=./proto \
+  --cpp_out=./server/cpp \
+  ./proto/helloworld.proto
