+++ proto/helloworld.proto
@@ -1,7 +1,8 @@
 syntax = "proto3";
 
 package helloworld;
-option go_package = "server/proto";
+
+option go_package = "proto/";
 
 service Greeter {
   rpc SayHello (HelloRequest) returns (HelloReply) {}
