+++ server/cpp/main.cpp
@@ -1,0 +1,35 @@
+#include <iostream>
+#include <memory>
+#include <string>
+
+#include <grpcpp/grpcpp.h>
+#include "helloworld.grpc.pb.h"
+
+using grpc::Server;
+using grpc::ServerBuilder;
+using grpc::ServerContext;
+using grpc::Status;
+using helloworld::Greeter;
+using helloworld::HelloReply;
+using helloworld::HelloRequest;
+
+class GreeterServiceImpl final : public Greeter::Service {
+    Status SayHello(ServerContext* context, const HelloRequest* request,
+                    HelloReply* reply) override {
+        reply->set_message("Hello " + request->name());
+        return Status::OK;
+    }
+};
+
+int main(int argc, char** argv) {
+    std::string server_address("0.0.0.0:50051");
+    GreeterServiceImpl service;
+
+    ServerBuilder builder;
+    builder.AddListeningPort(server_address, grpc::InsecureServerCredentials());
+    builder.RegisterService(&service);
+    std::unique_ptr<Server> server(builder.BuildAndStart());
+    std::cout << "gRPC server listening on " << server_address << std::endl;
+    server->Wait();
+    return 0;
+}
