--- server/images/README.md
@@ -1,1 +1,0 @@
-Place image files in this directory to make them accessible via the ImageService RPCs.
